+++ src/main/java/com/sparta/plate/service/product/ProductService.java
@@ -1,8 +1,7 @@
-package com.sparta.plate.service;
+package com.sparta.plate.service.product;
 
 import com.sparta.plate.dto.request.ProductRequestDto;
 import com.sparta.plate.entity.Product;
-import com.sparta.plate.entity.ProductHistory;
 import com.sparta.plate.entity.ProductImage;
 import com.sparta.plate.repository.ProductHistoryRepository;
 import com.sparta.plate.repository.ProductImageRepository;
@@ -43,14 +42,6 @@ public class ProductService {
         return savedProduct.getId();
     }
 
-    private UUID generateUniqueProductId() {
-        UUID productId = UUID.randomUUID();
-        while (productRepository.existsByProductId(productId)) {
-            productId = UUID.randomUUID();
-        }
-        return productId;
-    }
-
     @Transactional
     public void deleteProduct(UUID productId, Long userId) {
         Product product = productRepository.findById(productId).orElse(null);
@@ -62,31 +53,19 @@ public class ProductService {
     }
 
     @Transactional
-    public void deleteProductHistory(UUID historyId, Long userId) {
-        ProductHistory history = historyRepository.findById(historyId).orElse(null);
-
-        if (history != null) {
-            history.markAsDeleted(userId);
-            historyRepository.save(history);
-        }
-    }
-
-    @Transactional
-    public void deleteProductImage(UUID imageId, Long userId) {
-        ProductImage image = imageRepository.findById(imageId).orElse(null);
-
-        if (image != null) {
-            image.markAsDeleted(userId);
-            imageRepository.save(image);
-        }
-    }
-
-    @Transactional
     public void updateStockAndLimit(UUID productId, Long userId) {
         Product product = productRepository.findById(productId).orElse(null);
 
         if (product != null) {
             
         }
+    }
+
+    private UUID generateUniqueProductId() {
+        UUID productId = UUID.randomUUID();
+        while (productRepository.existsByProductId(productId)) {
+            productId = UUID.randomUUID();
+        }
+        return productId;
     }
 }
